requirements.txt
absl-py==2.2.2
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cvzone==1.6.1
cycler==0.12.1
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
keras==3.9.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mysqlclient==2.2.4
namex==0.0.8
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pillow==11.1.0
protobuf==4.25.6
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rich==14.0.0
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
SQLAlchemy==2.0.40
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.0.1
wrapt==1.17.2
zipp==3.21.0
